# [JavaCallableWrappers] Parallel.ForEach per assembly