@jonathanpeppers 2018-09-04
See bf0b5a0, an adventure in showing the issue with this idea...